misc: moved TaskTypes to folder
